+++ src/main/java/com/example/projectfirst/controller/UserController.java
@@ -1,12 +1,13 @@
 package com.example.projectfirst.controller;
 
-import com.example.projectfirst.entity.UserEntity;
+import com.example.projectfirst.entity.User;
+import com.example.projectfirst.exception.UserNotFoundException;
 import com.example.projectfirst.repository.UserRepo;
 import com.example.projectfirst.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/users")
@@ -19,27 +20,42 @@ public class UserController {
     }
 
     @GetMapping
-    public List<UserEntity> getAllUsers(){
+    public List<User> getAllUsers(){
         return userService.getAllUsers();
     }
 
     @PostMapping
-    public UserEntity createUser(@RequestBody UserEntity newUser){
-        return userService.saveOneUser(newUser);
+    public User createUser(@RequestBody User newUser){
+        return userService.createOneUser(newUser);
     }
 
     @GetMapping("/{userId}")
-    public UserEntity getOneUser(@PathVariable Long userId){
-        return userService.getOneUser(userId);
+    public User getOneUser(@PathVariable Long userId){
+        User user = userService.getOneUserById(userId);
+        if (user == null)
+            throw new UserNotFoundException();
+        return userService.getOneUserById(userId);
     }
 
     @PutMapping("{userId}")
-    public UserEntity updateOneUser(@PathVariable Long userId, @RequestBody UserEntity newUser){
-        return userService.updateOneUser(userId, newUser);
+    public User updateOneUser(@PathVariable Long userId, @RequestBody User newUser){
+        User user = userService.getOneUserById(userId);
+        if (user == null)
+            throw new UserNotFoundException();
+        return userService.updateOneUserById(userId, newUser);
     }
 
     @DeleteMapping("/{userId}")
     public void deleteOneUser(@PathVariable Long userId){
+        User user = userService.getOneUserById(userId);
+        if (user == null)
+            throw new UserNotFoundException();
         userService.deleteOneUser(userId);
+    }
+
+    @ExceptionHandler(UserNotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    private void handleUserNotFound(){
+
     }
 }
